+++ src/app/Playground/HåvardSamples/HåvardBottomSheet.xaml.cs
@@ -1,0 +1,48 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using DIPS.Mobile.UI.Components.BottomSheets;
+
+namespace Playground.HåvardSamples;
+
+public partial class HåvardBottomSheet : BottomSheet
+{
+    public HåvardBottomSheet()
+    {
+        InitializeComponent();
+    }
+
+    private void MenuItem_OnClicked(object sender, EventArgs e)
+    {
+        this.Close();
+    }
+
+    private void CloseAllBottomSheets(object sender, EventArgs e)
+    {
+        BottomSheetService.CloseAll();
+    }
+
+    private void Button_OnClicked(object sender, EventArgs e)
+    {
+        new BottomSheet()
+        {
+            Title = "Second bottom sheet",
+            Content = new Button() {Text = "Close all", Command = new Command(() => BottomSheetService.CloseAll())}
+        }.Open();
+    }
+
+    private void Switch_OnToggled(object sender, ToggledEventArgs e)
+    {
+        if (Positioning == Positioning.Fit)
+        {
+            Positioning = Positioning.Medium;
+        }
+        else
+        {
+            Positioning = Positioning.Fit;
+
+        }
+    }
+}